requirements.txt
attrs==22.2.0
click==8.1.3
coverage==7.2.1
flask[async]==2.2.3
Flask-SQLAlchemy==3.0.3
flask-swagger==0.2.14
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
PyMySQL==1.0.2
PyMySQL[rsa]==1.0.2
pluggy==1.0.0
pulsar-client==3.0.0
py==1.11.0
PyDispatcher==2.0.7
pyparsing==3.0.9
pytest==7.2.2
PyYAML==6.0
SQLAlchemy==2.0.4
tomli==2.0.1
Werkzeug==2.2.3
